fix: type
1. travelHistory type 추가
2. tarvelHistory type에 있는 것들을 import 해서 사용
3. props 변경
#11
