+++ src/components/atomic/molecules/TravelHistorySummary/index.tsx
@@ -3,16 +3,11 @@ import React from 'react';
 import Icon from '@/components/atomic/atoms/Icon';
 import Text from '@/components/atomic/atoms/Text';
 
+import { ITravelHistorySummary } from '@/types/travelHistory';
+
 import { Wrapper, DescriptionBox, FlexWrapper, ParticipantsCountWrapper } from './style';
 
-export interface ITravelHistorySummary {
-  title: string;
-  travelPeriod: string;
-  location: string;
-  participantsCount: number;
-}
-
-const TravelHistorySummary = ({ title, travelPeriod, location, participantsCount }: ITravelHistorySummary) => {
+const TravelHistorySummary = ({ title, startDate, endDate, location, participantsCount }: ITravelHistorySummary) => {
   return (
     <Wrapper>
       <Text fontColor="#1B1B1B" fontSize="15px" fontWeight="500">
@@ -20,7 +15,7 @@ const TravelHistorySummary = ({ title, travelPeriod, location, participantsCount
       </Text>
       <DescriptionBox>
         <Text fontColor="#707070" fontSize="11px" fontWeight="400">
-          {travelPeriod}
+          {startDate} - {endDate}
         </Text>
         <FlexWrapper>
           <Text fontColor="#707070" fontSize="11px" fontWeight="400">
