+++ src/components/atomic/organisms/TravelHistory/index.tsx
@@ -3,29 +3,18 @@ import React from 'react';
 import TravelHistoryImage from '@/components/atomic/molecules/TravelHistoryImage';
 import TravelHistorySummary from '@/components/atomic/molecules/TravelHistorySummary';
 
+import { ITravelHistory } from '@/types/travelHistory';
+
 import { Wrapper } from './style';
 
-import { ITravelHistorySummary } from '@/components/atomic/molecules/TravelHistorySummary';
-
-interface ITravelHistory {
-  src: string;
-  alt: string;
-}
-
-const TravelHistory = ({
-  src,
-  alt,
-  title,
-  travelPeriod,
-  location,
-  participantsCount,
-}: ITravelHistory & ITravelHistorySummary) => {
+const TravelHistory = ({ src, alt, title, startDate, endDate, location, participantsCount }: ITravelHistory) => {
   return (
     <Wrapper>
       <TravelHistoryImage src={src} alt={alt} />
       <TravelHistorySummary
         title={title}
-        travelPeriod={travelPeriod}
+        startDate={startDate}
+        endDate={endDate}
         location={location}
         participantsCount={participantsCount}
       />
